+++ app/src/API/Interfaces/SqlAction/FileAction.php
@@ -15,7 +15,7 @@ interface FileAction
     public function getFilename();
 
     /**
-     * The returned function should accept a single parameter, of type string
+     * The returned function should accept two parameters: execution-result and executor
      * @return callable|null
      */
     public function getResultsFilterCallable();