+++ app/src/API/Interfaces/SqlExecutor/Forked/FileExecutor.php
@@ -2,12 +2,13 @@
 
 namespace Db3v4l\API\Interfaces\SqlExecutor\Forked;
 
+use Db3v4l\API\Interfaces\SqlExecutor\Executor;
 use Symfony\Component\Process\Process;
 
 /**
  * Executes the SQL commands found in a file, using a Symfony Process
  */
-interface FileExecutor
+interface FileExecutor extends Executor
 {
     /**
      * @param string $filename