+++ app/src/API/Interfaces/SqlExecutor/InProcess/CommandExecutor.php
@@ -2,10 +2,12 @@
 
 namespace Db3v4l\API\Interfaces\SqlExecutor\InProcess;
 
+use Db3v4l\API\Interfaces\SqlExecutor\Executor;
+
 /**
  * Executes a SQL 'command' (ie. one or more statements) using php
  */
-interface CommandExecutor
+interface CommandExecutor extends Executor
 {
     /**
      * @param string $sql