+++ app/src/API/Interfaces/SqlExecutor/InProcess/FileExecutor.php
@@ -2,10 +2,12 @@
 
 namespace Db3v4l\API\Interfaces\SqlExecutor\InProcess;
 
+use Db3v4l\API\Interfaces\SqlExecutor\Executor;
+
 /**
  * Executes the SQL commands found in a file, using php
  */
-interface FileExecutor
+interface FileExecutor extends Executor
 {
     /**
      * @param string $filename