+++ app/src/Core/SqlExecutor/InProcess/FileExecutor.php
@@ -2,8 +2,8 @@
 
 namespace Db3v4l\Core\SqlExecutor\InProcess;
 
-use Db3v4l\API\Interfaces\InProcessCommandExecutor;
-use Db3v4l\API\Interfaces\InProcessFileExecutor;
+use Db3v4l\API\Interfaces\SqlExecutor\InProcess\CommandExecutor as InProcessCommandExecutor;
+use Db3v4l\API\Interfaces\SqlExecutor\InProcess\FileExecutor as InProcessFileExecutor;
 
 class FileExecutor implements InProcessFileExecutor
 {
@@ -18,7 +18,6 @@ class FileExecutor implements InProcessFileExecutor
         $this->wrappedExecutor = $wrappedExecutor;
     }
 
-
     /**
      * @param string $filename
      * @return Callable
@@ -26,5 +25,10 @@ class FileExecutor implements InProcessFileExecutor
     public function getExecuteFileCallable($filename)
     {
         return $this->wrappedExecutor->getExecuteCommandCallable(file_get_contents($filename));
+    }
+
+    public function resultSetToArray($data)
+    {
+        return $this->wrappedExecutor->resultSetToArray($data);
     }
 }