+++ app/src/Core/SqlExecutor/InProcess/PDO.php
@@ -2,16 +2,18 @@
 
 namespace Db3v4l\Core\SqlExecutor\InProcess;
 
-use Db3v4l\API\Interfaces\SqlExecutor\InProcess\StatementExecutor;
+use Db3v4l\API\Interfaces\SqlExecutor\InProcess\CommandExecutor;
 use Db3v4l\Core\SqlExecutor\BaseExecutor;
 
-class PDO extends BaseExecutor implements StatementExecutor
+class PDO extends BaseExecutor implements CommandExecutor
 {
-    /**
-     * @param string $sql
-     * @return Callable
-     */
     public function getExecuteCommandCallable($sql)
+    {
+        /// @todo
+        throw new \Exception('Not implemented yet');
+    }
+
+    public function resultSetToArray($data)
     {
         /// @todo
         throw new \Exception('Not implemented yet');